Move models to more appropriate
